Added ordered power off and on
In some cases there are dependencies on the order that
VMs must be shut down, backed up, then powered back on.
This commit adds the VM_SHUTDOWN_ORDER and
VM_STARTUP_ORDER configuration options which contain
a comma separated list of VM names in the order they must
be shut down and correspondingly started up.