Add additional external rotation
In some cases you may have to back up VMs to a temporary
location, and after the VMs are powered back on, move the
backups to the final location.  This patch enables file
rotation at a secondary directory, to limit the amount
of data that has to be transferred when doing an rsync
between the temporary and final location.