Enable chmod of backup files
Adds an option for the configuration file "BACKUP_FILES_CHMOD" for performing a chmod on the backup files. This is specifically intended to ease the use of a non-priviledged user to rsync the backup files; but may have other uses.

This can be a numerical value (755) or symbolic (UG+R,O-W).
Note that this value is applied recursively to the VMs current backup directory, so the value applies to the directories and files therein.