Fix creation of -symlink
Fix the creation of the -symlink link which can be useful when doing an rsync of the backup files. An existing -symlink entry is removed prior to the new one being created. This is to ensure that a new symlink is not merely added into the directory pointed to by the existing symlink.