FIX: Summary log when existing snapshot is found
Log to summary log when existing snapshot is found. Was not being logged due to the continue in the while loop for all VMs.

Also, capture summary end time in summary_logger function; we were capturing the end time just before calling the summary_logger, so have moved this within the function.