Ensure period for SMTP is on a line of its own
This change obeys the SMTP RFC closer and will allow more SMTP servers to work correctly. Current code has the "ghettoVCB LOG END" ending with \n, then in buildheaders it puts the SMTP "." right after that, making it "\n.\r\n". Some servers don't see this as a new line. It should be "\r\n.\r\n" so it is for sure on a new line.