Made addition to line 781
In your if statement on line 781, the ESX_RELEASE check is only for ESXi versions 5.5 and 6.0.  There is no check for 6.5.  When ran against a 6.5 host, it reverts to else statement concluding in the script being unable to run because it has the wrong command for mounting an NFS share.